# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FO76 DB",
    "description": "Локальный API утилиты FO76 DB: каталог предметов, инвентарь, роллы, события, настройки.\n\n**Доступ.** С этого компьютера (127.0.0.1, без прокси) токен не нужен. Из сети, если в настройках задан токен, передавайте\n`Authorization: Bearer <токен>` (или cookie `fo76db_token`, или `?token=` один раз). Без токена — 401.\nИзменяющие запросы с чужого `Origin` без Bearer отклоняются (защита от CSRF). Для браузерного фронтенда на другом адресе\nдобавьте его в `cors_origins` в `config.toml`. Запись путей, настроек сети, конфига IOM и подобное — только с этого компьютера.",
    "version": "2026.10.5"
  },
  "paths": {
...
    "/api/mule": {
      "post": {
        "summary": "Api Mule",
        "description": "Отметить персонажа складом (ему не предлагаются схемы) — без изменения имени и цвета.",
        "operationId": "api_mule_api_mule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_mule_api_mule_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "mule"
        ]
      }
    },
    "/api/known": {
      "post": {
        "summary": "Api Known",
        "operationId": "api_known_api_known_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_known_api_known_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "known"
        ]
      },
      "get": {
        "summary": "Api Known List",
        "description": "Изученные схемы персонажа: {FormID: auto|manual}.",
        "operationId": "api_known_list_api_known_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "known"
        ]
      }
    },
    "/api/known/bulk": {
      "post": {
        "summary": "Api Known Bulk",
        "operationId": "api_known_bulk_api_known_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_known_bulk_api_known_bulk_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "known"
        ]
      }
    },
    "/api/updates/plans": {
      "get": {
        "summary": "Api Updates Plans",
        "description": "Сколько схем и рецептов каждого обновления изучено у персонажа.",
        "operationId": "api_updates_plans_api_updates_plans_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "updates"
        ]
      }
    },
    "/api/optimize": {
      "get": {
        "summary": "Api Optimize",
        "operationId": "api_optimize_api_optimize_get",
        "parameters": [
          {
            "name": "scrapbox",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Scrapbox"
            }
          },
          {
            "name": "ammobox",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Ammobox"
            }
          },
          {
            "name": "heavy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10.0,
              "title": "Heavy"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "optimize"
        ]
      }
    },
    "/api/rolls": {
      "get": {
        "summary": "Api Rolls",
        "description": "Хотелки с числом совпадений и все легендарные предметы с оценкой.",
        "operationId": "api_rolls_api_rolls_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "rolls"
        ]
      },
      "post": {
        "summary": "Api Roll Save",
        "operationId": "api_roll_save_api_rolls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "rolls"
        ]
      }
    },
    "/api/rolls/{want_id}": {
      "delete": {
        "summary": "Api Roll Delete",
        "operationId": "api_roll_delete_api_rolls__want_id__delete",
        "parameters": [
          {
            "name": "want_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Want Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "rolls"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_known_api_known_post": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "formid": {
            "type": "string",
            "title": "Formid"
          },
          "known": {
            "type": "boolean",
            "title": "Known"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "formid",
          "known"
        ],
        "title": "Body_api_known_api_known_post"
      },
      "Body_api_known_bulk_api_known_bulk_post": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "formids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formids"
          },
          "known": {
            "type": "boolean",
            "title": "Known"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "formids",
          "known"
        ],
        "title": "Body_api_known_bulk_api_known_bulk_post"
      },
      "Body_api_mule_api_mule_post": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "mule": {
            "type": "boolean",
            "title": "Mule"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "mule"
        ],
        "title": "Body_api_mule_api_mule_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "token из config.toml / «Настройки → Доступ по сети»"
      }
    }
  },
...
}